feat: User Level과 모임 피드백 점수를 Double로 변경

+++ suldaksuldak-admin/src/main/java/com/sulsul/suldaksuldak/controller/admin/AdminUserController.java
@@ -29,7 +29,7 @@ public class AdminUserController {
             @ApiImplicitParam(name = "userPriKey", value = "유저 기본키", required = true, dataTypeClass = Long.class),
             @ApiImplicitParam(name = "nickname", value = "유저 닉네임", dataTypeClass = String.class),
             @ApiImplicitParam(name = "selfIntroduction", value = "유저 자기소개", dataTypeClass = String.class),
-            @ApiImplicitParam(name = "level", value = "유저 레벨", dataTypeClass = Integer.class),
+            @ApiImplicitParam(name = "level", value = "유저 레벨", dataTypeClass = Double.class),
             @ApiImplicitParam(name = "warningCnt", value = "신고 누적 횟수", dataTypeClass = Integer.class),
             @ApiImplicitParam(name = "isActive", value = "탈퇴 여부", dataTypeClass = Boolean.class)
     })
@@ -38,7 +38,7 @@ public class AdminUserController {
             Long userPriKey,
             String nickname,
             String selfIntroduction,
-            Integer level,
+            Double level,
             Integer warningCnt,
             Boolean isActive
     ) {
